pyproject.toml
[project]
name = "home-temperature"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-timescaledb>=0.2.13",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
]
